Fix pxe files missing during rollback
This change [1] copies the pxe files to a folder however in the rollback
scenario the file don't exist. This commit adds a condition if the folder
exists before copy.

Test Plan:

PASS: DX rollback after host-done state.

1: https://review.opendev.org/c/starlingx/update/+/938922

Story: 2010676
Task: 51557

+++ software/software/software_functions.py
@@ -1657,8 +1657,9 @@ def copy_pxeboot_cfg_files(to_major_release):
     src_dir = constants.FEED_DIR + "/rel-%s/pxeboot/pxelinux.cfg.files/" % to_major_release
     dst_dir = "/var/pxeboot/pxelinux.cfg.files"
     try:
-        shutil.copytree(src_dir, dst_dir, dirs_exist_ok=True)
-        LOG.info("Copied %s to %s" % (src_dir, dst_dir))
+        if os.path.exists(src_dir):
+            shutil.copytree(src_dir, dst_dir, dirs_exist_ok=True)
+            LOG.info("Copied %s to %s" % (src_dir, dst_dir))
     except Exception:
         LOG.exception("Error copying files from %s to: %s" % (src_dir, dst_dir))
         raise
